Fix: セッション関連のメソッドをヘルパーに切り出し
